GPII-1872: Split the implementation of kettle.dataSource.URL into its own file to make easier to use it in-browser, with a conversion of necessary node module dependencies with browserify.
